+++ javai-vector/src/main/java/dev/xtrafe/javai/vector/EmbeddingProviderTextEmbeddingsInference.java
@@ -112,4 +112,9 @@ public final class EmbeddingProviderTextEmbeddingsInference implements JavAIEmbe
             super(message, cause);
         }
     }
+
+    @Override
+    public String modelId() {
+        return modelId;
+    }
 }